done with the js
This code assumes that you have a global cart array that stores the items added to the cart. When the "minus" button is clicked, it checks if the quantity is greater than zero and decrements it if so. When the "plus" button is clicked, it simply increments the quantity. When the item itself is clicked (assuming you want to add it to the cart), it checks if the quantity is greater than zero and creates a new item object with the dish name, price, and quantity, and adds it to the cart array. It also logs a message to the console to confirm the item has been added, as well as the current total of the cart. Finally, there is a getCartTotal function that calculates the total price of all items in the cart, which you can use to display the cart total to the user.